+++ components/collector/tests/source_collectors/file_source_collectors/test_jacoco.py
@@ -11,6 +11,7 @@ class JaCoCoTest(SourceCollectorTestCase):
     """Unit tests for the JaCoCo metrics."""
 
     def setUp(self):
+        """Extend to set up a common source for the tests."""
         super().setUp()
         self.sources = dict(source_id=dict(type="jacoco", parameters=dict(url="https://jacoco/")))
 
@@ -18,15 +19,25 @@ class JaCoCoTest(SourceCollectorTestCase):
         """Test that the number of uncovered lines and the total number of lines are returned."""
         metric = dict(type="uncovered_lines", sources=self.sources, addition="sum")
         response = await self.collect(
-            metric, get_request_text="<report><counter type='LINE' missed='2' covered='4'/></report>")
+            metric, get_request_text="<report><counter type='LINE' missed='2' covered='4'/></report>"
+        )
         self.assert_measurement(response, value="2", total="6")
 
     async def test_uncovered_branches(self):
         """Test that the number of uncovered branches is returned."""
         metric = dict(type="uncovered_branches", sources=self.sources, addition="sum")
         response = await self.collect(
-            metric, get_request_text="<report><counter type='BRANCH' missed='4' covered='6'/></report>")
+            metric, get_request_text="<report><counter type='BRANCH' missed='4' covered='6'/></report>"
+        )
         self.assert_measurement(response, value="4", total="10")
+
+    async def test_uncovered_branches_without_branches(self):
+        """Test that a JaCoCo XML without branches results in 100% coverage."""
+        metric = dict(type="uncovered_branches", sources=self.sources, addition="sum")
+        response = await self.collect(
+            metric, get_request_text="<report><counter type='LINE' missed='4' covered='6'/></report>"
+        )
+        self.assert_measurement(response, value="0", total="0")
 
     async def test_source_up_to_dateness(self):
         """Test that the source age in days is returned."""
@@ -42,7 +53,8 @@ class JaCoCoTest(SourceCollectorTestCase):
         bytes_io = io.BytesIO()
         with zipfile.ZipFile(bytes_io, mode="w") as zipped_jacoco_report:
             zipped_jacoco_report.writestr(
-                "jacoco.xml", "<report><counter type='LINE' missed='2' covered='4'/></report>")
+                "jacoco.xml", "<report><counter type='LINE' missed='2' covered='4'/></report>"
+            )
         response = await self.collect(metric, get_request_content=bytes_io.getvalue())
         self.assert_measurement(response, value="2", total="6")
 
@@ -53,6 +65,7 @@ class JaCoCoTest(SourceCollectorTestCase):
         bytes_io = io.BytesIO()
         with zipfile.ZipFile(bytes_io, mode="w") as zipped_jacoco_report:
             zipped_jacoco_report.writestr(
-                "jacoco.html", "<html><body><p>Oops, user included the HTML instead of the XML</p></body></html>")
+                "jacoco.html", "<html><body><p>Oops, user included the HTML instead of the XML</p></body></html>"
+            )
         response = await self.collect(metric, get_request_content=bytes_io.getvalue())
         self.assert_measurement(response, value=None, connection_error="Zipfile contains no files with extension xml")